requirements.txt
django
djangorestframework
djangorestframework-xml
djoser
pillow
python-dotenv
django-filter
pytz
django-cors-headers
psycopg2-binary
whitenoise